Address review feedback, fix bugs
